--- Assets/JeongJH/CameraController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3b87377aa1af25e4ba812b22ee413f9d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
