+++ Assets/JeongJH/Material.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a313da0e08cb5654a95944249a4d0bba
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
