+++ Assets/JeongJH/Material/New Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 02c90ed0579fb6948b410b57ab76bebe
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
