--- Assets/JeongJH/NPCTrigger.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d393d2f84dee222419dd161efa00eb06
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
