--- Assets/JeongJH/Outline Shader.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 56ea228e5c69be64d88f9e39e8214d74
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
