+++ Assets/JeongJH/Prefabs/Slope.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b9a4f5e5573de75469d021d233975d37
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
