+++ Assets/JeongJH/Prefabs/SpawnedStoneCircle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cf836ea6c49dbae43aae5fde810e0e55
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
