--- Assets/JeongJH/Robbery.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d9320b7688f5afd4182c4905afed6000
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
