--- Assets/JeongJH/TransParentCubeMat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 172016c729fa905448a0ef4045ceac53
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
