--- Assets/JeongJH/Zero Friction.physicMaterial
@@ -1,14 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!134 &13400000
-PhysicMaterial:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_Name: Zero Friction
-  dynamicFriction: 0
-  staticFriction: 0
-  bounciness: 1
-  frictionCombine: 0
-  bounceCombine: 3
