--- Assets/JeongJH/Zero Friction.physicMaterial.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 93bbe7b4aafbde143b6a8a46b199a61c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 13400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
