--- Assets/JeongJH/scrips.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5134e2c8a39a5944bd337ef0e9aeeac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
